createdb.py: Made DB-agnostic (between MySQL and Postgres) Signed-off-by: Martin Stone <martin@d7415.co.uk>

+++ createdb.py
@@ -27,6 +27,8 @@ from Core.config import Config
 from Core.db import Base, session
 import shipstats
 
+mysql = Config.get("DB", "dbms") == "mysql"
+
 if len(sys.argv) > 2 and sys.argv[1] == "--migrate":
     round = sys.argv[2]
     if round.isdigit():
@@ -39,33 +41,33 @@ else:
     print "To migrate from an old round use: createdb.py --migrate <previous_round>"
     sys.exit()
 
-# if round:
-#     print "Moving tables to '%s' schema"%(round,)
-#     try:
-#         session.execute(text("ALTER SCHEMA public RENAME TO %s;" % (round,)))
-#     except ProgrammingError:
-#         print "Oops! Either you don't have permission to modify schemas or you already have a backup called '%s'" % (round,)
-#         session.rollback()
-#         sys.exit()
-#     else:
-#         session.commit()
-#     finally:
-#         session.close()
+if round and not mysql:
+    print "Moving tables to '%s' schema"%(round,)
+    try:
+        session.execute(text("ALTER SCHEMA public RENAME TO %s;" % (round,)))
+    except ProgrammingError:
+        print "Oops! Either you don't have permission to modify schemas or you already have a backup called '%s'" % (round,)
+        session.rollback()
+        sys.exit()
+    else:
+        session.commit()
+    finally:
+        session.close()
 
 print "Importing database models"
 from Core.maps import Channel
-"""
-print "Creating schema and tables"
-try:
-    session.execute(text("CREATE SCHEMA public;"))
-except ProgrammingError:
-    print "A public schema already exists, but this is completely normal"
-    session.rollback()
-else:
-    session.commit()
-finally:
-    session.close()
-"""
+if not mysql:
+    print "Creating schema and tables"
+    try:
+        session.execute(text("CREATE SCHEMA public;"))
+    except ProgrammingError:
+        print "A public schema already exists, but this is completely normal"
+        session.rollback()
+    else:
+        session.commit()
+    finally:
+        session.close()
+
 Base.metadata.create_all()
 
 print "Setting up default channels"
@@ -91,42 +93,42 @@ for chan, name in Config.items("Channels"):
         session.commit()
 session.close()
 
-# if round:
-#     print "Migrating data:"
-#     try:
-#         print "  - users/friends"
-#         session.execute(text("INSERT INTO users (id, name, alias, passwd, active, access, url, email, phone, pubphone, _smsmode, sponsor, quits, available_cookies, carebears, last_cookie_date, fleetcount) SELECT id, name, alias, passwd, active, access, url, email, phone, pubphone, _smsmode::varchar::smsmode, sponsor, quits, available_cookies, carebears, last_cookie_date, 0 FROM %s.users;" % (round,)))
-#         session.execute(text("SELECT setval('users_id_seq',(SELECT max(id) FROM users));"))
-#         session.execute(text("INSERT INTO phonefriends (user_id, friend_id) SELECT user_id, friend_id FROM %s.phonefriends;" % (round,)))
-#         print "  - slogans/quotes"
-#         session.execute(text("INSERT INTO slogans (text) SELECT text FROM %s.slogans;" % (round,)))
-#         session.execute(text("INSERT INTO quotes (text) SELECT text FROM %s.quotes;" % (round,)))
-#         print "  - props/votes/cookies"
-#         session.execute(text("INSERT INTO invite_proposal (id,active,proposer_id,person,created,closed,vote_result,comment_text) SELECT id,active,proposer_id,person,created,closed,vote_result,comment_text FROM %s.invite_proposal;" % (round,)))
-#         session.execute(text("INSERT INTO kick_proposal (id,active,proposer_id,person_id,created,closed,vote_result,comment_text) SELECT id,active,proposer_id,person_id,created,closed,vote_result,comment_text FROM %s.kick_proposal;" % (round,)))
-#         session.execute(text("SELECT setval('proposal_id_seq',(SELECT max(id) FROM (SELECT id FROM invite_proposal UNION SELECT id FROM kick_proposal) AS proposals));"))
-#         session.execute(text("INSERT INTO prop_vote (vote,carebears,prop_id,voter_id) SELECT vote,carebears,prop_id,voter_id FROM %s.prop_vote;" % (round,)))
-#         session.execute(text("INSERT INTO cookie_log (log_time,year,week,howmany,giver_id,receiver_id) SELECT log_time,year,week,howmany,giver_id,receiver_id FROM %s.cookie_log;" % (round,)))
-#         print "  - smslog"
-#         session.execute(text("INSERT INTO sms_log (sender_id,receiver_id,phone,sms_text,mode) SELECT sender_id,receiver_id,phone,sms_text,mode FROM %s.sms_log;" % (round,)))
-#     except DBAPIError, e:
-#         print "An error occurred during migration: %s" %(str(e),)
-#         session.rollback()
-#         print "Reverting to previous schema"
-#         """session.execute(text("DROP SCHEMA public CASCADE;"))
-#         session.execute(text("ALTER SCHEMA %s RENAME TO public;" % (round,)))"""
-#         session.commit()
-#         sys.exit()
-#     else:
-#         session.commit()
-#     finally:
-#         session.close()
-"""
-if round == "temp":
-    print "Deleting temporary schema"
-    session.execute(text("DROP SCHEMA temp CASCADE;"))
-    session.commit()
-    session.close()
-"""
+if round and not mysql:
+    print "Migrating data:"
+    try:
+        print "  - users/friends"
+        session.execute(text("INSERT INTO users (id, name, alias, passwd, active, access, url, email, phone, pubphone, _smsmode, sponsor, quits, available_cookies, carebears, last_cookie_date, fleetcount) SELECT id, name, alias, passwd, active, access, url, email, phone, pubphone, _smsmode::varchar::smsmode, sponsor, quits, available_cookies, carebears, last_cookie_date, 0 FROM %s.users;" % (round,)))
+        session.execute(text("SELECT setval('users_id_seq',(SELECT max(id) FROM users));"))
+        session.execute(text("INSERT INTO phonefriends (user_id, friend_id) SELECT user_id, friend_id FROM %s.phonefriends;" % (round,)))
+        print "  - slogans/quotes"
+        session.execute(text("INSERT INTO slogans (text) SELECT text FROM %s.slogans;" % (round,)))
+        session.execute(text("INSERT INTO quotes (text) SELECT text FROM %s.quotes;" % (round,)))
+        print "  - props/votes/cookies"
+        session.execute(text("INSERT INTO invite_proposal (id,active,proposer_id,person,created,closed,vote_result,comment_text) SELECT id,active,proposer_id,person,created,closed,vote_result,comment_text FROM %s.invite_proposal;" % (round,)))
+        session.execute(text("INSERT INTO kick_proposal (id,active,proposer_id,person_id,created,closed,vote_result,comment_text) SELECT id,active,proposer_id,person_id,created,closed,vote_result,comment_text FROM %s.kick_proposal;" % (round,)))
+        session.execute(text("SELECT setval('proposal_id_seq',(SELECT max(id) FROM (SELECT id FROM invite_proposal UNION SELECT id FROM kick_proposal) AS proposals));"))
+        session.execute(text("INSERT INTO prop_vote (vote,carebears,prop_id,voter_id) SELECT vote,carebears,prop_id,voter_id FROM %s.prop_vote;" % (round,)))
+        session.execute(text("INSERT INTO cookie_log (log_time,year,week,howmany,giver_id,receiver_id) SELECT log_time,year,week,howmany,giver_id,receiver_id FROM %s.cookie_log;" % (round,)))
+        print "  - smslog"
+        session.execute(text("INSERT INTO sms_log (sender_id,receiver_id,phone,sms_text,mode) SELECT sender_id,receiver_id,phone,sms_text,mode FROM %s.sms_log;" % (round,)))
+    except DBAPIError, e:
+        print "An error occurred during migration: %s" %(str(e),)
+        session.rollback()
+        print "Reverting to previous schema"
+        """session.execute(text("DROP SCHEMA public CASCADE;"))
+        session.execute(text("ALTER SCHEMA %s RENAME TO public;" % (round,)))"""
+        session.commit()
+        sys.exit()
+    else:
+        session.commit()
+    finally:
+        session.close()
+
+    if round == "temp":
+        print "Deleting temporary schema"
+        session.execute(text("DROP SCHEMA temp CASCADE;"))
+        session.commit()
+        session.close()
+
 print "Inserting ship stats"
 shipstats.main()
